+++ types/ErrorMessageId.ts
@@ -1,0 +1,7 @@
+enum ErrorMessageId {
+  InvalidEmail = 'invalid-email',
+  RequiredField = 'required-field',
+  InvalidField = 'invalid-field',
+}
+
+export default ErrorMessageId;
