+++ types/FieldRule.ts
@@ -1,0 +1,15 @@
+import ErrorMessageId from './ErrorMessageId';
+
+export enum RuleType {
+  Required = 'required',
+  Regex = 'regex',
+}
+
+type RuleValue = string | boolean;
+type FieldRule = {
+  type: RuleType;
+  value: RuleValue;
+  errorId?: ErrorMessageId;
+};
+
+export default FieldRule;
